Cargo updated. Changed the physics components to use f64. Added an input_config. Re-added the player module. Started refactoring `PlayerControlSystem` to use amethyst. Removed an bunch of stuff from `PlayerControlSystem` because they were obseleted by amethyst. Moved configs from resources to assets/configs. Copy pasted a spritesheet loading function from zesty green and made it less hardcoded.

+++ src/physics.rs
@@ -4,13 +4,13 @@ use amethyst::ecs::{Component, DenseVecStorage, ParJoin, ReadStorage, System, Ve
 
 #[derive(Debug, Component)]
 #[storage(VecStorage)]
-pub struct Position(pub Point2<f32>);
+pub struct Position(pub Point2<f64>);
 
 #[derive(Debug, Component)]
-pub struct Velocity(pub Vector2<f32>);
+pub struct Velocity(pub Vector2<f64>);
 
 #[derive(Debug, Component)]
-pub struct Acceleration(pub Vector2<f32>);
+pub struct Acceleration(pub Vector2<f64>);
 
 pub struct PhysicsSystem;
